refactor(devtokens): consolidate developer token code in plugin
